tests/submission/W3C/domhtml/HTMLObjectElement12.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-object-element.html#dom-object-usemap" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object data="/support/images/yellow-rectangle.png" height="50" type="image/png" usemap="#shapes" width="100"></object>

<map name="shapes">
<area shape=rect coords="0,0,100,50" href="/support/html/sample.html" alt="HTML test link">
</map>

<script>
var domtest = async_test("The useMap IDL attribute must reflect the usemap content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vusemap;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      vusemap = testNode.useMap;

      assert_equals(vusemap,"#shapes");
       
    domtest.done();
}));

</script>
<p>The useMap IDL attribute must reflect the usemap content attribute.</p>
<div id="log"></div>
</body></html>
